+++ src/include/fn_stream.h
@@ -1,0 +1,332 @@
+/*
+ *
+ * MIT License
+ *
+ * Copyright (C) 2019 YaweiZhang <yawei.zhang@foxmail.com>.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * ===============================================================================
+ *
+ * (end of COPYRIGHT)
+ */
+
+
+ /*
+  * AUTHORS:  YaweiZhang <yawei.zhang@foxmail.com>
+  * VERSION:  1.0.0
+  * PURPOSE:  fn-log is a cpp-based logging utility.
+  * CREATION: 2019.4.20
+  * RELEASED: 2019.6.27
+  * QQGROUP:  524700770
+  */
+
+
+#pragma once
+#ifndef _FN_LOG_STREAM_H_
+#define _FN_LOG_STREAM_H_
+
+#include "fn_data.h"
+#include "fn_parse.h"
+#include "fn_load.h"
+#include "fn_core.h"
+
+namespace FNLog
+{
+    class LogStream
+    {
+    public:
+        static const int MAX_CONTAINER_DEPTH = 5;
+    public:
+        explicit LogStream(LogStream&& ls) noexcept
+        {
+            logger_ = ls.logger_;
+            log_data_ = ls.log_data_;
+            ls.logger_ = nullptr;
+            ls.log_data_ = nullptr;
+        }
+
+        explicit LogStream(Logger& logger, int channel_id, int priority, int category, 
+            const char * const file_name, int file_name_len, int line,
+            const char * const func_name, int func_name_len, unsigned int prefix)
+        {
+            logger_ = nullptr;
+            log_data_ = nullptr;
+            if (CanPushLog(logger, channel_id, priority, category) != 0)
+            {
+                return;
+            }
+            logger_ = &logger;
+            log_data_ = AllocLogData(logger, channel_id, priority, category, prefix);
+            if (prefix == LOG_PREFIX_NULL)
+            {
+                return;
+            }
+            if (prefix & LOG_PREFIX_FILE)
+            {
+                write_char_unsafe(' ');
+                if (file_name && file_name_len > 0)
+                {
+                    int jump_bytes = short_path(file_name, file_name_len);
+                    write_buffer_unsafe(file_name + jump_bytes, file_name_len - jump_bytes);
+                }
+                else
+                {
+                    write_buffer_unsafe("nofile", 6);
+                }
+                write_char_unsafe(':');
+                write_char_unsafe('<');
+                *this << (unsigned long long)line;
+                write_char_unsafe('>');
+                write_char_unsafe(' ');
+            }
+            if (prefix & LOG_PREFIX_FUNCTION)
+            {
+                if (func_name && func_name_len > 0)
+                {
+                    write_buffer_unsafe(func_name, func_name_len);
+                }
+                else
+                {
+                    write_buffer_unsafe("null", 4);
+                }
+                write_char_unsafe(' ');
+            }
+        }
+        
+        ~LogStream()
+        {
+            if (log_data_) 
+            {
+                PushLog(*logger_, log_data_);
+                log_data_ = nullptr;
+                logger_ = nullptr;
+            }
+        }
+        
+        LogStream& set_category(int category) { if (log_data_) log_data_->category_ = category;  return *this; }
+        LogStream& write_char_unsafe(char ch)
+        {
+            log_data_->content_[log_data_->content_len_] = ch;
+            log_data_->content_len_++;
+            return *this;
+        }
+        LogStream& write_buffer_unsafe(const char* src, int src_len)
+        {
+            memcpy(log_data_->content_ + log_data_->content_len_, src, src_len);
+            log_data_->content_len_ += src_len;
+            return *this;
+        }
+
+        LogStream& write_buffer(const char* src, int src_len)
+        {
+            if (log_data_ && src && src_len > 0 && log_data_->content_len_ < LogData::MAX_LOG_SIZE)
+            {
+                src_len = FN_MIN(src_len, LogData::MAX_LOG_SIZE - log_data_->content_len_);
+                memcpy(log_data_->content_ + log_data_->content_len_, src, src_len);
+                log_data_->content_len_ += src_len;
+            }
+            return *this;
+        }
+
+        LogStream& write_pointer(const void* ptr)
+        {
+            if (log_data_ && log_data_->content_len_ + 30 <= LogData::MAX_LOG_SIZE)
+            {
+                log_data_->content_len_ += FNLog::write_pointer_unsafe(log_data_->content_ + log_data_->content_len_, ptr);
+            }
+            return *this;
+        }
+
+        LogStream& write_binary(const char* dst, int len)
+        {
+            if (!log_data_)
+            {
+                return *this;
+            }
+            write_buffer("\r\n\t[", sizeof("\r\n\t[")-1);
+            for (int i = 0; i < (len / 32) + 1; i++)
+            {
+                write_buffer("\r\n\t[", sizeof("\r\n\t[") - 1);
+                *this << (void*)(dst + (size_t)i * 32);
+                write_buffer(": ", sizeof(": ") - 1);
+                for (int j = i * 32; j < (i + 1) * 32 && j < len; j++)
+                {
+                    if (isprint((unsigned char)dst[j]))
+                    {
+                        write_buffer(" ", sizeof(" ") - 1);
+                        write_buffer(dst + j, 1);
+                        write_buffer(" ", sizeof(" ") - 1);
+                    }
+                    else
+                    {
+                        write_buffer(" . ", sizeof(" . ") - 1);
+                    }
+                }
+                write_buffer("\r\n\t[", sizeof("\r\n\t[") - 1);
+                if (log_data_->content_len_ + sizeof(void*) <= LogData::MAX_LOG_SIZE)
+                {
+                    write_pointer(dst + (size_t)i * 32);
+                }
+                write_buffer(": ", sizeof(": ") - 1);
+                for (int j = i * 32; j < (i + 1) * 32 && j < len; j++)
+                {
+                    if (log_data_->content_len_ + 30 >= LogData::MAX_LOG_SIZE)
+                    {
+                        break;
+                    }
+                    log_data_->content_len_ += FNLog::write_hex_unsafe<2>(log_data_->content_ + log_data_->content_len_,
+                        (unsigned long long)(unsigned char)dst[j]);
+                    write_buffer(" ", sizeof(" ") - 1);
+                }
+            }
+            write_buffer("\r\n\t]\r\n\t", sizeof("\r\n\t]\r\n\t") - 1);
+            return *this;
+        }
+
+        LogStream& operator <<(const char* cstr)
+        {
+            if (log_data_)
+            {
+                if (cstr)
+                {
+                    write_buffer(cstr, (int)strlen(cstr));
+                }
+                else
+                {
+                    write_buffer("null", 4);
+                }
+            }
+            return *this;
+        }
+        LogStream& operator <<(const void* ptr)
+        {
+            write_pointer(ptr);
+            return *this;
+        }
+        
+        LogStream& operator <<(std::nullptr_t) 
+        {
+            return write_pointer(nullptr);
+            return *this;
+        }
+
+        LogStream& operator << (char ch) { return write_buffer(&ch, 1);}
+        LogStream & operator << (unsigned char ch) { return (*this << (unsigned long long)ch); }
+
+        LogStream& operator << (bool val) { return (val ? write_buffer("true", 4) : write_buffer("false", 5)); }
+
+        LogStream & operator << (short val) { return (*this << (long long)val); }
+        LogStream & operator << (unsigned short val) { return (*this << (unsigned long long)val); }
+        LogStream & operator << (int val) { return (*this << (long long)val); }
+        LogStream & operator << (unsigned int val) { return (*this << (unsigned long long)val); }
+        LogStream & operator << (long val) { return (*this << (long long)val); }
+        LogStream & operator << (unsigned long val) { return (*this << (unsigned long long)val); }
+        
+        LogStream& operator << (long long integer)
+        {
+            if (log_data_ && log_data_->content_len_ + 30 <= LogData::MAX_LOG_SIZE)
+            {
+                log_data_->content_len_ += write_dec_unsafe<0>(log_data_->content_ + log_data_->content_len_, (long long)integer);
+            }
+            return *this;
+        }
+
+        LogStream& operator << (unsigned long long integer)
+        {
+            if (log_data_ && log_data_->content_len_ + 30 <= LogData::MAX_LOG_SIZE)
+            {
+                log_data_->content_len_ += write_dec_unsafe<0>(log_data_->content_ + log_data_->content_len_, (unsigned long long)integer);
+            }
+            return *this;
+        }
+
+        LogStream& operator << (float f)
+        {
+            if (log_data_ && log_data_->content_len_ + 30 <= LogData::MAX_LOG_SIZE)
+            {
+                log_data_->content_len_ += write_float_unsafe(log_data_->content_ + log_data_->content_len_, f);
+            }
+            return *this;
+        }
+        LogStream& operator << (double df)
+        {
+            if (log_data_ && log_data_->content_len_ + 30 <= LogData::MAX_LOG_SIZE)
+            {
+                log_data_->content_len_ += write_double_unsafe(log_data_->content_ + log_data_->content_len_, df);
+            }
+            return *this;
+        }
+
+
+        template<class _Ty1, class _Ty2>
+        inline LogStream & operator <<(const std::pair<_Ty1, _Ty2> & val){ return *this << '<' <<val.first << ':' << val.second << '>'; }
+
+        template<class Container>
+        LogStream& write_container(const Container& container, const char* name, int len)
+        {
+            write_buffer(name, len);
+            write_buffer("(", 1);
+            *this << container.size();
+            write_buffer(")[", 2);
+            int inputCount = 0;
+            for (auto iter = container.begin(); iter != container.end(); iter++)
+            {
+                if (inputCount >= MAX_CONTAINER_DEPTH)
+                {
+                    *this << "..., ";
+                    break;
+                }
+                if(inputCount > 0)
+                {
+                    *this << ", ";
+                }
+                *this << *iter;
+                inputCount++;
+            }
+            return *this << "]";
+        }
+
+        template<class _Elem, class _Alloc>
+        LogStream & operator <<(const std::vector<_Elem, _Alloc> & val) { return write_container(val, "vector:", sizeof("vector:") - 1);}
+        template<class _Elem, class _Alloc>
+        LogStream & operator <<(const std::list<_Elem, _Alloc> & val) { return write_container(val, "list:", sizeof("list:") - 1);}
+        template<class _Elem, class _Alloc>
+        LogStream & operator <<(const std::deque<_Elem, _Alloc> & val) { return write_container(val, "deque:", sizeof("deque:") - 1);}
+        template<class _Key, class _Tp, class _Compare, class _Allocator>
+        LogStream & operator <<(const std::map<_Key, _Tp, _Compare, _Allocator> & val) { return write_container(val, "map:", sizeof("map:") - 1);}
+        template<class _Key, class _Compare, class _Allocator>
+        LogStream & operator <<(const std::set<_Key, _Compare, _Allocator> & val) { return write_container(val, "set:", sizeof("set:") - 1);}
+        template<class _Key, class _Tp, class _Hash, class _Compare, class _Allocator>
+        LogStream& operator <<(const std::unordered_map<_Key, _Tp, _Hash, _Compare, _Allocator>& val)
+        {return write_container(val, "unordered_map:", sizeof("unordered_map:") - 1);}
+        template<class _Key, class _Hash, class _Compare, class _Allocator>
+        LogStream& operator <<(const std::unordered_set<_Key, _Hash, _Compare, _Allocator> & val)
+        {return write_container(val, "unordered_set:", sizeof("unordered_set:") - 1);}
+        template<class _Traits, class _Allocator>
+        LogStream & operator <<(const std::basic_string<char, _Traits, _Allocator> & str) { return write_buffer(str.c_str(), (int)str.length());}
+        
+    public:
+        LogData * log_data_ = nullptr;
+        Logger* logger_ = nullptr;
+    };
+}
+
+
+#endif
