+++ tests/fast_start_debug_test.cpp
@@ -34,7 +34,16 @@ int main(int argc, char* argv[])
     LOGFMTA("now time:<%llu>", (unsigned long long)time(nullptr));
     LOGFMTF("now time:<%llu>", (unsigned long long)time(nullptr));
 
-    LogDebugPack(0, 0, "now time:", time(nullptr), ";");
+
+    PackTrace("now time:", time(nullptr), ";");
+    PackDebug("now time:", time(nullptr), ";");
+    PackInfo("now time:", time(nullptr), ";");
+    PackWarn("now time:", time(nullptr), ";");
+    PackError("now time:", time(nullptr), ";");
+    PackAlarm("now time:", time(nullptr), ";");
+    PackFatal("now time:", time(nullptr), ";");
+
+
     
     LogAlarm() << "finish";
